backend/requirements.txt
fastapi
uvicorn
librosa
numpy
python-multipart
yt-dlp
pydantic
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
beautifulsoup4
lxml
psycopg2-binary
pgvector
asyncpg
